Timeout value aproximation & new std baud
Timeout is now uint8_t and use internal clock aproximation for faster
computation, yields to a bigger timeout.
Baud on confi struct is now uint32_t so to accept bigger bauds. Std baud
changed to 250k .

+++ RosaSensors/RosaSensors/Rs485Std.cpp
@@ -26,7 +26,7 @@ namespace Rosa
 		
 	}
 
-	bool Rs485Std::read_msg( uint8_t& data_length, uint8_t *msg, float byte_timeout, float response_timeout)
+	bool Rs485Std::read_msg( uint8_t& data_length, uint8_t *msg, uint8_t byte_timeout, uint8_t response_timeout)
 	{
 		rx_enable->unset();
